requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-countries==7.5.1
django-cuser==2017.3.16
django-debug-toolbar==4.1.0
django-filter==23.2
django-rest-swagger==2.2.0
django-tinymce==3.6.1
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
modeltranslation==0.25
numpy==1.24.3
openapi-codec==1.3.2
packaging==23.1
pandas==2.0.2
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
